+++ packages/koop-cache-memory/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* Replace highland with node stream Readable
+
 ## [1.1.0] - 2017-02-15
 ### Added
 * Upsert API